monitoring-demo/app/requirements.txt
fastapi
uvicorn
prometheus-client
prometheus-fastapi-instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging